requirements.txt
requests
pytest
pytest-html
allure-pytest
pyyaml
loguru
openpyxl
pymysql
redis
cryptography
jsonpath